+++ src/type-guards.ts
@@ -1,21 +1,29 @@
-import {IMixinBase, IUseMixins, Mixin, MixinAssertError} from "./types";
-import {AnyObject, Constructor} from "./common.types";
+import {AnyObject, CheckNever, Constructor, IMixinBase, IUseMixins, Mixin, MixinAssertError, MixinsPropObject} from "./types";
 
 // noinspection JSUnusedLocalSymbols
-export function haveMixin<M extends Mixin<any, any>, Class extends AnyObject = {}>(
-    v: AnyObject, mixin: M,
+export function haveMixin<M extends Mixin<any, any>, Class extends AnyObject | any = never>(
+    v: AnyObject,
+    mixin: M,
     vClass?: Class
-): v is IUseMixins<[M], Class> {
+): v is (
+    CheckNever<Class,
+        IUseMixins<Class, M>,
+        MixinsPropObject<M>
+>) {
     return typeof v === "object" && 'mixins' in v && typeof v.mixins === 'object' &&
         mixin.mixinName in v.mixins && typeof v.mixins[mixin.mixinName] === 'object';
 }
 
 // noinspection JSUnusedLocalSymbols
-export function haveMixins<Mixins extends Array<Mixin<any, any>>, Class extends AnyObject = {}>(
+export function haveMixins<Mixins extends Array<Mixin<any, any>>, Class extends AnyObject | any = never>(
     v: AnyObject,
     mixins: Mixins,
     vClass?: Class,
-): v is IUseMixins<Mixins, Class> {
+): v is (
+    CheckNever<Class,
+        IUseMixins<Class, Mixins[0], Mixins[1], Mixins[2], Mixins[3], Mixins[4]>,
+        MixinsPropObject<Mixins[0], Mixins[1], Mixins[2], Mixins[3], Mixins[4]>
+>) {
     const baseCheck = typeof v === "object" && 'mixins' in v && typeof v.mixins === 'object';
 
     if (!baseCheck) {
@@ -23,7 +31,7 @@ export function haveMixins<Mixins extends Array<Mixin<any, any>>, Class extends 
     }
 
     for (const mixin of mixins) {
-        if (haveMixin(v, mixin) === false) {
+        if (haveMixin(v, mixin, vClass) === false) {
             return false;
         }
     }
@@ -42,12 +50,16 @@ export function isMixin<M extends IMixinBase<any> = IMixinBase<any>>(value: any)
 
 /// asserts functions
 
-export function assertHaveMixin<M extends Mixin<any, any>, Class extends AnyObject = {}>(
+export function assertHaveMixin<M extends Mixin<any, any>, Class extends AnyObject | any = never>(
     v: AnyObject,
     mixin: M,
     vClass?: Class,
     error?: string | Error,
-): asserts v is IUseMixins<[M], typeof v> {
+): asserts v is (
+    CheckNever<Class,
+        IUseMixins<Class, M>,
+        MixinsPropObject<M>
+>) {
     if (haveMixin(v, mixin, vClass) === false) {
         if (error instanceof Error) {
             throw error;
@@ -58,12 +70,17 @@ export function assertHaveMixin<M extends Mixin<any, any>, Class extends AnyObje
 }
 
 
-export function assertHaveMixins<Mixins extends Array<Mixin<any, any>>, Class extends AnyObject = {}>(
+
+export function assertHaveMixins<MM extends Array<Mixin<any, any>>, Class extends AnyObject | any = never>(
     v: AnyObject,
-    mixins: Mixins,
+    mixins: MM,
     vClass?: Class,
     error?: string | Error,
-): asserts v is IUseMixins<Mixins, typeof v> {
+): asserts v is (
+    CheckNever<Class,
+        IUseMixins<Class, MM[0], MM[1], MM[2], MM[3], MM[4], MM[5], MM[6], MM[7], MM[8], MM[9]>,
+        MixinsPropObject<MM[0], MM[1], MM[2], MM[3], MM[4], MM[5], MM[6], MM[7], MM[8], MM[9]>
+>) {
     if (haveMixins(v, mixins, vClass) === false) {
         if (error instanceof Error) {
             throw error;